+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 **Added**
 
+- #166 Better handling of ClientPatientID and Patient fields in Add Sample
 
 **Changed**